Enhance pagination and validation in API and UI components: implement paginated fetching for shifts, visits, and vitals, update frontend hooks and API endpoints to support pagination, and integrate Zod validation for improved data handling in forms. Additionally, add validation error handling in registration and login pages for better user experience.

+++ backend/src/migrations/create_refresh_tokens_table.sql
@@ -1,7 +1,6 @@
--- リフレッシュトークンテーブルを作成
 CREATE TABLE IF NOT EXISTS refresh_tokens (
-  id INT AUTO_INCREMENT PRIMARY KEY,
-  user_id INT NOT NULL,
+  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  user_id BIGINT NOT NULL,
   token VARCHAR(255) NOT NULL UNIQUE,
   expires_at DATETIME NOT NULL,
   revoked TINYINT(1) NOT NULL DEFAULT 0,
@@ -11,4 +10,3 @@ CREATE TABLE IF NOT EXISTS refresh_tokens (
   INDEX idx_expires_at (expires_at),
   FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
-
